requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
Pillow>=10.0.0
pillow-heif>=0.14.0
imagehash>=4.3.0
pydantic>=2.0.0
